==> rmd/1c_merging_arrest_charge_data.Rmd <==
---
title: "1c_Combining Arrest Data"
author: "Brit Henderson"
date: "Last Updated: `r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

# Objective
Combine the different charge types into one arrest counts file
```{r directories}
readdatdir <- "/data/share/xproject/Training/Practice/henderson/Dissertation/rda/Intermediate"
rdadir <- "/data/share/xproject/Training/Practice/henderson/Dissertation/rda"

```

```{r functions}

```

```{r packages}
# # Import packages
library(Hmisc)
library(tidyverse) # suite of packages in for manipulating/tidying/visualizing, etc. data
library(lubridate) # manipulating dates
library(readxl)  # read in excel files
library(knitr) # for RMD formatting
library(kableExtra) # to output tables as images
library(fastDummies) # easy creation of dummy variables 



```


#Load the data sets
```{r data}
dat.index <- readRDS(here::here("rda/Intermediate/int_index.rds"))
dat.prop <- readRDS(here::here("rda/Intermediate/int_prop.rds"))
dat.drug <- readRDS(here::here("rda/Intermediate/int_drug.rds"))
dat.other <- readRDS(here::here("rda/Intermediate/int_other.rds"))




```

# merge data sets
```{r merge}

#put all data frames into list
df_list <- list(dat.index, dat.prop, dat.drug, dat.other)      

#merge all data frames together
all.arrests <- df_list %>% reduce(full_join, 
                                  by= c("ori","ori9","year","state","country_division", "fips_place_code","agency_name","fips_state_county_code", "state_abb",
                                        "population_group","population","longitude","latitude","fips_county_code", "fips_state_code","covered_by","number_of_months_reported"))


checkdat1 <- select(all.arrests, agency_name, year, state, fips_state_county_code) %>% group_by(agency_name, year, state, fips_state_county_code) %>% 
  summarise(numrec = n())

table(checkdat1$numrec, useNA = "always")

checkdat2 <- filter(checkdat1, numrec >1)

#Removing exact duplicate observations
all.arrests2 <- all.arrests[!duplicated(all.arrests), ]


#Removing partial duplicates for agencies that contain multiple records for same year. In these 27 instances, one record has a population of 0
checkdat1 <- all.arrests2 %>% 
  filter( ori9 !="0") %>% 
  group_by(ori9) %>% 
  summarise(years_included = n())

n_distinct(ayears$ori9)

extrayears <- as.list(filter(ayears, years_included > 46) %>% select(ori9))


checkdat2 <- all.arrests2[all.arrests2$ori9 %in% extrayears[[1]] ,]
n_distinct(checkdat2$ori9)
#11 agencies have multiple records for the same year, need to delete 27 records
checkdat3 <- filter(checkdat2, population==0)

all.arrests3 <- filter(all.arrests2, !(ori9 %in% extrayears[[1]]) | (ori9 %in% extrayears[[1]] & population != 0))


```



#summary statistics for all arrests data set
```{r stats}
summary(all.arrests3)
describe(all.arrests3)
glimpse(all.arrests3)

checkdat4 <- all.arrests3 %>% 
  group_by(year, agency_name, fips_place_code, fips_state_county_code) %>% 
  summarise(n = n())

summary(checkdat4)

hist(checkdat4$n)
table(checkdat4$n, useNA = "always")
checkdat4 <- all.arrests3 %>% count(year, agency_name, fips_place_code, fips_state_county_code)
```

# Create measure of total cases
```{r}
all.arrests4 <- all.arrests3 %>% 
  mutate(young_adult_tot = young_adult_drug_tot + young_adult_index_tot + young_adult_prop_tot + young_adult_other_tot,
            other_adult_tot = other_adult_drug_tot + other_adult_prop_tot + other_adult_index_tot + other_adult_other_tot,
            all_adult_tot = all_adult_drug_tot + all_adult_index_tot + all_adult_prop_tot + all_adult_other_tot)

#looks good
```

#Aggregate to County level
```{r}
county.arrests <- all.arrests4 %>% 
  group_by(year,state,country_division, fips_state_county_code, state_abb,fips_county_code, fips_state_code) %>% 
  mutate(min_number_of_months_reported = min(number_of_months_reported),
            c_young_adult_prop_tot = sum(young_adult_prop_tot),
            c_young_adult_drug_tot = sum(young_adult_drug_tot),
            c_young_adult_index_tot = sum(young_adult_index_tot),
            c_young_adult_other_tot = sum(young_adult_other_tot),
            c_other_adult_drug_tot = sum(other_adult_drug_tot),
            c_other_adult_prop_tot = sum(other_adult_prop_tot),
            c_other_adult_index_tot = sum(other_adult_index_tot),
            c_other_adult_other_tot = sum(other_adult_index_tot),
            c_all_adult_drug_tot = sum(all_adult_drug_tot),
            c_all_adult_prop_tot = sum(all_adult_prop_tot),
            c_all_adult_index_tot = sum(all_adult_index_tot),
            c_all_adult_other_tot = sum(all_adult_other_tot),
            c_young_adult_tot = sum(young_adult_tot),
            c_other_adult_tot = sum(other_adult_tot),
            c_all_adult_tot = sum(all_adult_tot))
```

```{r}
checkdat5 <- county.arrests %>% filter(year==2000, fips_state_county_code=="23017") %>% 
  select(year, fips_state_county_code, agency_name, state, contains("young") )
#looks good!
```

# Keep only one record per county-year with only total arrests
```{r}
final_county <-  county.arrests %>% distinct(year, fips_state_county_code, state,c_young_adult_tot,c_other_adult_tot,c_all_adult_tot)
```

#Save files to type rda

```{r}
saveRDS(all.arrests2, file = file.path(rdadir, "all_arrests"))
saveRDS(all.arrests4, file = file.path(rdadir, "arrests_by_agency.rds"))
saveRDS(final_county, file = file.path(rdadir, "total_arrests_by_county.rds"))
```
